+++ packages/salesforcedx-vscode-lightning/src/constants.ts
@@ -1,8 +1,0 @@
-/*
- * Copyright (c) 2019, salesforce.com, inc.
- * All rights reserved.
- * Licensed under the BSD 3-Clause license.
- * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-export const COMPONENT_LIBRARY_BUNDLE_LINK =
-  'https://developer.salesforce.com/docs/component-library/bundle/';