+++ packages/salesforcedx-vscode-lwc/src/commands/forceLightningLwcStart.ts
@@ -35,6 +35,16 @@ const SfdxCommandletExecutor = sfdxCoreExports.SfdxCommandletExecutor;
 const logName = 'force_lightning_lwc_start';
 const commandName = nls.localize(`force_lightning_lwc_start_text`);
 
+/**
+ * Hints for providing a user-friendly error message / action.
+ * Hints come from the stderr output of lwc-dev-server. (We should move this to lwc-dev-server later)
+ */
+export const enum errorHints {
+  SERVER_STARTUP_FALIED = 'Server start up failed',
+  ADDRESS_IN_USE = 'EADDRINUSE',
+  INACTIVE_SCRATCH_ORG = 'Error authenticating to your scratch org. Make sure that it is still active'
+}
+
 export interface ForceLightningLwcStartOptions {
   /** whether to automatically open the browser after server start */
   openBrowser: boolean;
@@ -44,6 +54,7 @@ export interface ForceLightningLwcStartOptions {
 
 export class ForceLightningLwcStartExecutor extends SfdxCommandletExecutor<{}> {
   private readonly options: ForceLightningLwcStartOptions;
+  private errorHint?: string;
 
   constructor(options: ForceLightningLwcStartOptions = { openBrowser: true }) {
     super();
@@ -84,6 +95,7 @@ export class ForceLightningLwcStartExecutor extends SfdxCommandletExecutor<{}> {
     channelService.showChannelOutput();
 
     let serverStarted = false;
+    let printedError = false;
 
     const progress = new Subject();
     ProgressNotification.show(
@@ -114,25 +126,40 @@ export class ForceLightningLwcStartExecutor extends SfdxCommandletExecutor<{}> {
       }
     });
 
-    // handler errors
-    execution.processExitSubject.subscribe(async exitCode => {
-      DevServerService.instance.clearServerHandler(serverHandler);
-
-      if (!serverStarted && !cancellationToken.isCancellationRequested) {
-        let message = nls.localize('force_lightning_lwc_start_failed');
-
-        // TODO proper exit codes in lwc-dev-server for address in use, auth/org error, etc.
-        if (exitCode === 127) {
-          message = nls.localize('force_lightning_lwc_start_not_found');
+    execution.stderrSubject.subscribe(async data => {
+      if (!printedError && data) {
+        let errorCode = -1;
+        if (data.toString().includes(errorHints.SERVER_STARTUP_FALIED)) {
+          errorCode = 1;
         }
+        if (data.toString().includes(errorHints.ADDRESS_IN_USE)) {
+          errorCode = 98;
+        }
+        if (data.toString().includes(errorHints.INACTIVE_SCRATCH_ORG)) {
+          this.errorHint = errorHints.INACTIVE_SCRATCH_ORG;
+        }
+        if (errorCode !== -1) {
+          this.handleErrors(
+            cancellationToken,
+            serverHandler,
+            serverStarted,
+            errorCode
+          );
+          progress.complete();
+          printedError = true;
+        }
+      }
+    });
 
-        showError(new Error(message), logName, commandName);
-      } else if (exitCode !== undefined && exitCode !== null && exitCode > 0) {
-        const message = nls.localize(
-          'force_lightning_lwc_start_exited',
+    execution.processExitSubject.subscribe(async exitCode => {
+      if (!printedError) {
+        this.handleErrors(
+          cancellationToken,
+          serverHandler,
+          serverStarted,
           exitCode
         );
-        showError(new Error(message), logName, commandName);
+        printedError = true;
       }
     });
 
@@ -147,6 +174,39 @@ export class ForceLightningLwcStartExecutor extends SfdxCommandletExecutor<{}> {
       );
       this.showChannelOutput();
     });
+  }
+
+  private handleErrors(
+    cancellationToken: vscode.CancellationToken,
+    serverHandler: ServerHandler,
+    serverStarted: boolean,
+    exitCode: number | null | undefined
+  ) {
+    DevServerService.instance.clearServerHandler(serverHandler);
+    if (!serverStarted && !cancellationToken.isCancellationRequested) {
+      let message = nls.localize('force_lightning_lwc_start_failed');
+
+      if (
+        exitCode === 1 &&
+        this.errorHint === errorHints.INACTIVE_SCRATCH_ORG
+      ) {
+        message = nls.localize('force_lightning_lwc_inactive_scratch_org');
+      }
+      if (exitCode === 127) {
+        message = nls.localize('force_lightning_lwc_start_not_found');
+      }
+      if (exitCode === 98) {
+        message = nls.localize('force_lightning_lwc_start_addr_in_use');
+      }
+
+      showError(new Error(message), logName, commandName);
+    } else if (exitCode !== undefined && exitCode !== null && exitCode > 0) {
+      const message = nls.localize(
+        'force_lightning_lwc_start_exited',
+        exitCode
+      );
+      showError(new Error(message), logName, commandName);
+    }
   }
 }