+++ packages/salesforcedx-vscode-lwc/test/vscode-integration/commands/forceLightningLwcStart.test.ts
@@ -18,6 +18,7 @@ import * as vscode from 'vscode';
 import { DEV_SERVER_BASE_URL } from '../../../src/commands/commandConstants';
 import * as commandUtils from '../../../src/commands/commandUtils';
 import {
+  errorHints,
   forceLightningLwcStart,
   ForceLightningLwcStartExecutor
 } from '../../../src/commands/forceLightningLwcStart';
@@ -192,6 +193,36 @@ describe('forceLightningLwcStart', () => {
         );
       });
 
+      it('shows the error message if server start up failed', () => {
+        const executor = new ForceLightningLwcStartExecutor();
+        const fakeExecution = new FakeExecution(executor.build());
+        cliCommandExecutorStub.returns(fakeExecution);
+
+        executor.execute({ type: 'CONTINUE', data: {} });
+
+        fakeExecution.stderrSubject.next(errorHints.SERVER_STARTUP_FALIED);
+        sinon.assert.notCalled(
+          notificationServiceStubs.showSuccessfulExecutionStub
+        );
+
+        sinon.assert.calledTwice(notificationServiceStubs.showErrorMessageStub);
+        sinon.assert.calledWith(
+          notificationServiceStubs.showErrorMessageStub,
+          sinon.match(
+            nls.localize(
+              'command_failure',
+              nls.localize(`force_lightning_lwc_start_text`)
+            )
+          )
+        );
+
+        sinon.assert.calledOnce(channelServiceStubs.appendLineStub);
+        sinon.assert.calledWith(
+          channelServiceStubs.appendLineStub,
+          sinon.match(nls.localize('force_lightning_lwc_start_failed'))
+        );
+      });
+
       it('opens the browser once server is started', () => {
         const executor = new ForceLightningLwcStartExecutor();
         const fakeExecution = new FakeExecution(executor.build());
@@ -229,6 +260,66 @@ describe('forceLightningLwcStart', () => {
           channelServiceStubs.appendLineStub,
           sinon.match(nls.localize('force_lightning_lwc_start_not_found'))
         );
+      });
+
+      it('shows an error when the address is already in use', () => {
+        const executor = new ForceLightningLwcStartExecutor();
+        const fakeExecution = new FakeExecution(executor.build());
+        cliCommandExecutorStub.returns(fakeExecution);
+
+        executor.execute({ type: 'CONTINUE', data: {} });
+        fakeExecution.processExitSubject.next(98);
+
+        const commandName = nls.localize(`force_lightning_lwc_start_text`);
+
+        sinon.assert.calledTwice(notificationServiceStubs.showErrorMessageStub);
+        sinon.assert.calledWith(
+          notificationServiceStubs.showErrorMessageStub,
+          sinon.match(nls.localize('command_failure', commandName))
+        );
+
+        sinon.assert.calledOnce(channelServiceStubs.appendLineStub);
+        sinon.assert.calledWith(
+          channelServiceStubs.appendLineStub,
+          sinon.match(nls.localize('force_lightning_lwc_start_addr_in_use'))
+        );
+      });
+
+      it('shows an error when scratch org is inactive', () => {
+        const executor = new ForceLightningLwcStartExecutor();
+        const fakeExecution = new FakeExecution(executor.build());
+        cliCommandExecutorStub.returns(fakeExecution);
+
+        executor.execute({ type: 'CONTINUE', data: {} });
+        fakeExecution.stderrSubject.next(errorHints.INACTIVE_SCRATCH_ORG);
+        fakeExecution.processExitSubject.next(1);
+
+        const commandName = nls.localize(`force_lightning_lwc_start_text`);
+
+        sinon.assert.calledTwice(notificationServiceStubs.showErrorMessageStub);
+        sinon.assert.calledWith(
+          notificationServiceStubs.showErrorMessageStub,
+          sinon.match(nls.localize('command_failure', commandName))
+        );
+
+        sinon.assert.calledOnce(channelServiceStubs.appendLineStub);
+        sinon.assert.calledWith(
+          channelServiceStubs.appendLineStub,
+          sinon.match(nls.localize('force_lightning_lwc_inactive_scratch_org'))
+        );
+      });
+
+      it('shows no error when server is stopping', () => {
+        const executor = new ForceLightningLwcStartExecutor();
+        const fakeExecution = new FakeExecution(executor.build());
+        cliCommandExecutorStub.returns(fakeExecution);
+
+        executor.execute({ type: 'CONTINUE', data: {} });
+        fakeExecution.stdoutSubject.next('Server up');
+        fakeExecution.processExitSubject.next(0);
+
+        sinon.assert.notCalled(notificationServiceStubs.showErrorMessageStub);
+        sinon.assert.notCalled(channelServiceStubs.appendLineStub);
       });
 
       it('shows an error message when the process exists before server startup', () => {